Changed u32 to i64

+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub fn rgb_to_ycbcr(r: u32, g: u32, b: u32) -> (u32, u32, u32) {
+pub fn rgb_to_ycbcr(r: i64, g: i64, b: i64) -> (i64, i64, i64) {
     // let y = 0.299 * r + 0.587 * g + 0.114 * b;
     // let cb = 128.0 - 0.1687 * r - 0.3313 * g + 0.5 * b;
     // let cr = 128.0 + 0.5 * r - 0.4187 * g - 0.0813 * b;
@@ -16,14 +16,14 @@ pub fn rgb_to_ycbcr(r: u32, g: u32, b: u32) -> (u32, u32, u32) {
     let r = r as f64;
     let g = g as f64;
     let b = b as f64;
-    let y = (0.0 + 0.299 * r + 0.587 * g + 0.114 * b) as u32;
-    let cb = (128.0 - 0.169 * r - 0.331 * g + 0.500 * b) as u32;
-    let cr = (128.0 + 0.500 * r - 0.419 * g - 0.081 * b) as u32;
+    let y = (0.0 + 0.299 * r + 0.587 * g + 0.114 * b) as i64;
+    let cb = (128.0 - 0.169 * r - 0.331 * g + 0.500 * b) as i64;
+    let cr = (128.0 + 0.500 * r - 0.419 * g - 0.081 * b) as i64;
 
     (y, cb, cr)
 }
 
-pub fn ycbcr_to_rgb(y: u32, cb: u32, cr: u32) -> (u32, u32, u32) {
+pub fn ycbcr_to_rgb(y: i64, cb: i64, cr: i64) -> (i64, i64, i64) {
     // https://stackoverflow.com/questions/4041840/function-to-convert-ycbcr-to-rgb
     // let r = (y as f64 + 1.40200 * (cr - 0x80) as f64) as u32;
     // let g = (y as f64 - 0.34414 * (cb - 0x80) as f64 - 0.71414 * (cr - 0x80) as f64) as u32;
@@ -33,9 +33,9 @@ pub fn ycbcr_to_rgb(y: u32, cb: u32, cr: u32) -> (u32, u32, u32) {
     let y = y as f64;
     let cb = cb as f64;
     let cr = cr as f64;
-    let r = (1.0 * y + 0.000 * (cb - 128.) + 1.400 * (cr - 128.)) as u32;
-    let g = (1.0 * y - 0.343 * (cb - 128.) - 0.711 * (cr - 128.)) as u32;
-    let b = (1.0 * y + 1.765 * (cb - 128.) + 0.000 * (cr - 128.)) as u32;
+    let r = (1.0 * y + 0.000 * (cb - 128.) + 1.400 * (cr - 128.)) as i64;
+    let g = (1.0 * y - 0.343 * (cb - 128.) - 0.711 * (cr - 128.)) as i64;
+    let b = (1.0 * y + 1.765 * (cb - 128.) + 0.000 * (cr - 128.)) as i64;
 
     (r, g, b)
 }
